requirements.txt
django-cors-headers==4.2.0
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.0.1
psycopg2-binary